Move command utils
